Fix CodeFactor
