+++ WalletWasabi/BitcoinCore/IRPCClient.cs
@@ -1,5 +1,4 @@
 ﻿using NBitcoin;
-using NBitcoin.RPC;
 using System.Threading.Tasks;
 
 namespace WalletWasabi.BitcoinCore
@@ -10,33 +9,5 @@ namespace WalletWasabi.BitcoinCore
 		Task<uint256> GetBestBlockHashAsync();
 		Task<Block> GetBlockAsync(uint256 blockId);
 		Task<BlockHeader> GetBlockHeaderAsync(uint256 blockHash);
-
-	}
-
-	public class RpcWrappedClient : IRPCClient
-	{
-		private RPCClient Rpc { get; }
-
-		public RpcWrappedClient(RPCClient rpc)
-		{
-			Rpc = rpc;
-		}
-		
-		public Network Network => Rpc.Network;
-
-		public Task<uint256> GetBestBlockHashAsync()
-		{
-			return Rpc.GetBestBlockHashAsync();
-		}
-
-		public Task<Block> GetBlockAsync(uint256 blockId)
-		{
-			return Rpc.GetBlockAsync(blockId);
-		}
-
-		public Task<BlockHeader> GetBlockHeaderAsync(uint256 blockHash)
-		{
-			return Rpc.GetBlockHeaderAsync(blockHash);
-		}
 	}
 }
